requirements.txt
psycopg2-binary
Jinja2
MarkupSafe
alembic
Mako
python-dateutil
SQLAlchemy-Utils
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.1.1
Werkzeug==2.3.6
WTForms==3.0.1
SQLAlchemy==2.0.17
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
email-validator==2.0.0
